+++ tests/ImageHelperTest.php
@@ -4,14 +4,14 @@ use \X\Util\ImageHelper;
 use \X\Util\FileHelper;
 
 const INPUT_DIR = __DIR__ . '/input';
-const INPUT_BACKUP_DIR = __DIR__ . '/input-backup';
-const OUTPUT_DIR = __DIR__ . '/input';
+const INPUT_BAK_DIR = __DIR__ . '/input-backup';
+const OUTPUT_DIR = __DIR__ . '/output';
 
 final class ImageHelperTest extends TestCase {
   public static function setUpBeforeClass(): void {
     // During testing, files in the input directory are overwritten, so reset the input directory before testing.
     FileHelper::delete(INPUT_DIR);
-    FileHelper::copyDirectory(INPUT_BACKUP_DIR, INPUT_DIR);
+    FileHelper::copyDirectory(INPUT_BAK_DIR, INPUT_DIR);
   }
 
   public function testWriteFirstFrameOfGifInSeparateFile(): void {
@@ -28,12 +28,12 @@ final class ImageHelperTest extends TestCase {
   }
 
   public function testGetNumberOfGifFrames(): void {
-    $inputPath = INPUT_DIR . '/sample2.gif';
+    $inputPath = INPUT_DIR . '/sample_2.gif';
     $this->assertSame(ImageHelper::getNumberOfGifFrames($inputPath), 19);
   }
 
   public function testGetNumberOfFramesInGifWithNoAnimation(): void {
-    $inputPath = INPUT_DIR . '/sample-non-animated.gif';
+    $inputPath = INPUT_DIR . '/non-animated.gif';
     $this->assertSame(ImageHelper::getNumberOfGifFrames($inputPath), 1);
   }
 }